// app/api/ussd/states/state.ts
export enum State {
  START = "start",
  REGISTRATION = "registration",
  REGISTERNAME = "registerName",
  REGISTERBUSINESS = "registerBusiness",
  REGISTERLOCATION = "registerLocation",
  REGISTRATIONEND = "registrationEnd",
  ABOUTUS = "aboutUs",
  ENDMENU = "endMenu",
  MAINMENU = "mainMenu",
  USERPROFILE = "userProfile",
  EDITBASICINFO = "editBasicInfo",
  EDITNAME = "editName",
  EDITBUSINESS = "editBusiness",
  EDITLOCATION = "editLocation",
  PROFILESAVED = "profileSaved",
  VIEWPROFILE = "viewProfile",
  RECORDKEEPING = "recordKeeping",
  LOGREVENUE = "logRevenue",
  REVENUEAMOUNT = "revenueAmount",
  REVENUEQUANTITY = "revenueQuantity",
  REVENUELOGGED = "revenueLogged",
  LOGEXPENSES = "logExpenses",
  EXPENSETYPE = "expenseType",
  EXPENSELOGGED = "expenseLogged",
  INVENTORYMANAGEMENT = "inventoryManagement",
  ADDSTOCK = "addStock",
  STOCKNAME = "stockName",
  STOCKAMOUNT = "stockAmount",
  STOCKUNITTYPE = "stockUnitType",
  STOCKUNITPRICE = "stockUnitPrice",
  STOCKADDED = "stockAdded",
  CHECKINVENTORYLEVELS = "checkInventoryLevels",
  BUSINESSHEALTHSUMMARY = "businessHealthSummary",
  VIEWPROFITLOSS = "viewProfitLoss",
  FINANCIALHEALTHSCORE = "financialHealthScore",
  FINANCIALGOALSANDTIPS = "financialGoalsAndTips",
  SETFINANCIALGOALS = "setFinancialGoals",
  GOALSET = "goalSet",
  SAVINGSADVICE = "savingsAdvice",
  FINANCIALTIPS = "financialTips",
  REMINDERS = "reminders",
  SETRECURRINGEXPENSEREMINDER = "setRecurringExpenseReminder",
  SETREMINDERINTERVAL = "setReminderInterval",
  REMINDERSET = "reminderSet",
  SETCUSTOMREMINDER = "setCustomReminder",
  EXPORTDATASUPPORT = "exportDataSupport",
  EXPORTDATA = "exportData",
  HELPANDSUPPORT = "helpAndSupport",
  LOANS = "loans",
  ELIGIBILITYCHECKER = "eligibilityChecker",
  LOANOVERVIEW = "loanOverview",
  EXIT = "exit",
}